test and batch drafting
